requirements.txt
autopep8
beautifulsoup4
boto3
botocore
celery
dj-database-url
Django
django-allauth
django-cors-headers
django-dbbackup
django-filter
django-redis
django-rest-auth
django-rest-swagger
django-silk
django-storages
django-templated-mail
django_debug_toolbar
djangorestframework
djangorestframework-simplejwt
djoser
drf-nested-routers
flower
gunicorn
Jinja2
mysqlclient
Pillow
pipenv
psycopg2
psycopg2-binary
Pygments
PyJWT
pyrsistent
PySocks
python-dateutil
python-decouple
python-dotenv
python-magic
pytz
pyxdg
PyYAML
QtPy
redis
reportlab
requests
requests-oauthlib
s3transfer
simplejson
six
social-auth-app-django
social-auth-core
virtualenv
whitenoise
django-ckeditor
django-admin-thumbnails
django-defender
django-recaptcha
channels
honcho
daphne
channels-redis